Update metafiles and tests structure
