+++ .gitattributes
@@ -1,10 +1,9 @@
-/.github            export-ignore
-/.gitattributes     export-ignore
-/.gitignore         export-ignore
-/phpunit.xml.dist   export-ignore
-/tests              export-ignore
-/.editorconfig      export-ignore
-/.php_cs.dist.php   export-ignore
-/psalm.xml          export-ignore
-/psalm.xml.dist     export-ignore
-/UPGRADING.md       export-ignore
+* text=auto
+
+.github                export-ignore
+.*                     export-ignore
+*.dist                 export-ignore
+*.yaml                 export-ignore
+*.xml                  export-ignore
+composer.lock          export-ignore
+tests                  export-ignore
