+++ .gitignore
@@ -9,3 +9,4 @@ docs
 vendor
 node_modules
 .php-cs-fixer.cache
+/runtime
