--- .styleci.yml
@@ -1,73 +1,0 @@
-preset: psr12
-risky: true
-
-version: 8
-
-enabled:
-  - alpha_ordered_traits
-  - array_indentation
-  - array_push
-  - combine_consecutive_issets
-  - combine_consecutive_unsets
-  - combine_nested_dirname
-  - declare_strict_types
-  - dir_constant
-  - fully_qualified_strict_types
-  - function_to_constant
-  - is_null
-  - magic_constant_casing
-  - magic_method_casing
-  - method_separation
-  - modernize_types_casting
-  - native_function_casing
-  - native_function_type_declaration_casing
-  - no_alias_functions
-  - no_empty_comment
-  - no_empty_phpdoc
-  - no_empty_statement
-  - no_extra_block_blank_lines
-  - no_short_bool_cast
-  - no_superfluous_elseif
-  - no_unneeded_control_parentheses
-  - no_unneeded_curly_braces
-  - no_unneeded_final_method
-  - no_unset_cast
-  - no_unused_imports
-  - no_unused_lambda_imports
-  - no_useless_else
-  - no_useless_return
-  - normalize_index_brace
-  - php_unit_dedicate_assert
-  - php_unit_dedicate_assert_internal_type
-  - php_unit_expectation
-  - php_unit_mock
-  - php_unit_mock_short_will_return
-  - php_unit_namespaced
-  - php_unit_no_expectation_annotation
-  - phpdoc_no_empty_return
-  - phpdoc_no_useless_inheritdoc
-  - phpdoc_order
-  - phpdoc_property
-  - phpdoc_scalar
-  - phpdoc_separation
-  - phpdoc_singular_inheritdoc
-  - phpdoc_trim
-  - phpdoc_trim_consecutive_blank_line_separation
-  - phpdoc_type_to_var
-  - phpdoc_types
-  - phpdoc_types_order
-  - print_to_echo
-  - regular_callable_call
-  - return_assignment
-  - self_accessor
-  - self_static_accessor
-  - set_type_to_cast
-  - short_array_syntax
-  - short_list_syntax
-  - simplified_if_return
-  - single_quote
-  - standardize_not_equals
-  - ternary_to_null_coalescing
-  - trailing_comma_in_multiline_array
-  - unalign_double_arrow
-  - unalign_equals
